fix clippy warning
