+++ crates/cairo-lint-core/src/lints/manual/manual_ok_or.rs
@@ -4,7 +4,7 @@ use cairo_lang_syntax::node::ast::ExprMatch;
 use cairo_lang_syntax::node::db::SyntaxGroup;
 use cairo_lang_syntax::node::TypedSyntaxNode;
 
-use crate::lints::manual::manual::{check_manual, ManualLint};
+use crate::lints::manual::{check_manual, ManualLint};
 
 pub const MANUAL_OK_OR: &str = "Manual match for Option<T> detected. Consider using ok_or instead";
